Broken forms

+++ getaway/webapp/src/main/java/ar/edu/itba/getaway/webapp/forms/ExperienceForm.java
@@ -15,7 +15,7 @@ public class  ExperienceForm {
     private String experienceName;
 
     @NotNull
-    private Long experienceCategory;
+    private Integer experienceCategory;
 
     @NotEmpty
     private String experienceCountry;
@@ -58,7 +58,7 @@ public class  ExperienceForm {
     public String getExperienceUrl() {
         return experienceUrl;
     }
-    public Long getExperienceCategory() {
+    public Integer getExperienceCategory() {
         return experienceCategory;
     }
     public MultipartFile getExperienceImg() {
@@ -79,7 +79,7 @@ public class  ExperienceForm {
     public void setExperienceAddress(String experienceAddress) {
         this.experienceAddress = experienceAddress;
     }
-    public void setExperienceCategory(Long experienceCategory) {
+    public void setExperienceCategory(Integer experienceCategory) {
         this.experienceCategory = experienceCategory;
     }
     public void setExperienceImg(MultipartFile experienceImg) {
